# openapi_spec.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Islamic Finance Workflows API",
        "description": "AI-powered AAOIFI-compliant document generation using Claude and Graphiti",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/documents/ingest": {
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Ingest Document",
                "description": "Upload and ingest a document into Graphiti.\n\nArgs:\n    file: PDF or DOCX file to upload\n    type: 'aaoifi' or 'context'\n\nReturns:\n    UploadedDocument with episode_id",
                "operationId": "ingest_document_api_documents_ingest_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_ingest_document_api_documents_ingest_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IngestDocumentResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/workflows/execute": {
            "post": {
                "tags": [
                    "Workflows"
                ],
                "summary": "Execute Workflow",
                "description": "Execute a workflow (placeholder).\n\nArgs:\n    request: ExecuteWorkflowRequest with template_id and configuration\n\nReturns:\n    execution_id for tracking",
                "operationId": "execute_workflow_api_workflows_execute_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExecuteWorkflowRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExecuteWorkflowResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/workflows/{execution_id}/stream": {
            "get": {
                "tags": [
                    "Workflows"
                ],
                "summary": "Stream Claude",
                "description": "Stream Claude execution (placeholder).\n\nArgs:\n    execution_id: Workflow execution ID\n\nReturns:\n    SSE stream of Claude responses",
                "operationId": "stream_claude_api_workflows__execution_id__stream_get",
                "parameters": [
                    {
                        "name": "execution_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Execution Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_ingest_document_api_documents_ingest_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File"
                    },
                    "type": {
                        "type": "string",
                        "title": "Type"
                    }
                },
                "type": "object",
                "required": [
                    "file",
                    "type"
                ],
                "title": "Body_ingest_document_api_documents_ingest_post"
            },
            "ExecuteWorkflowRequest": {
                "properties": {
                    "template_id": {
                        "type": "string",
                        "title": "Template Id"
                    },
                    "context_text": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Context Text"
                    },
                    "context_document_ids": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Context Document Ids"
                    },
                    "user_notes": {
                        "additionalProperties": {
                            "type": "string"
                        },
                        "type": "object",
                        "title": "User Notes"
                    }
                },
                "type": "object",
                "required": [
                    "template_id"
                ],
                "title": "ExecuteWorkflowRequest",
                "description": "Request to execute a workflow"
            },
            "ExecuteWorkflowResponse": {
                "properties": {
                    "execution_id": {
                        "type": "string",
                        "title": "Execution Id"
                    }
                },
                "type": "object",
                "required": [
                    "execution_id"
                ],
                "title": "ExecuteWorkflowResponse",
                "description": "Response from workflow execution start"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "IngestDocumentResponse": {
                "properties": {
                    "document": {
                        "$ref": "#/components/schemas/UploadedDocument"
                    }
                },
                "type": "object",
                "required": [
                    "document"
                ],
                "title": "IngestDocumentResponse",
                "description": "Response from document ingestion"
            },
...
            "UploadedDocument": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "filename": {
                        "type": "string",
                        "title": "Filename"
                    },
                    "filesize": {
                        "type": "integer",
                        "title": "Filesize"
                    },
                    "mime_type": {
                        "type": "string",
                        "title": "Mime Type"
                    },
                    "uploaded_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Uploaded At"
                    },
                    "episode_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Episode Id"
                    },
                    "parsed": {
                        "type": "boolean",
                        "title": "Parsed",
                        "default": false
                    },
                    "content_preview": {
                        "type": "string",
                        "title": "Content Preview",
                        "default": ""
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "filename",
                    "filesize",
                    "mime_type",
                    "uploaded_at"
                ],
                "title": "UploadedDocument",
                "description": "Document uploaded by user"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}